move out ticker css
