+++ assets/pages/new-contributions-landing/components/ContributionFormContainer.jsx
@@ -20,7 +20,7 @@ import { openDirectDebitPopUp } from 'components/directDebit/directDebitActions'
 
 import { type State } from '../contributionsLandingReducer';
 import { NewContributionForm } from './ContributionForm';
-import { ContributionTicker } from './ContributionTicker';
+import { ContributionTicker } from './ContributionTicker/ContributionTicker';
 import { setPayPalHasLoaded } from '../contributionsLandingActions';
 
 import {
