+++ assets/pages/new-contributions-landing/components/ContributionTicker/ContributionTicker.jsx
