+++ assets/pages/new-contributions-landing/components/ContributionTicker/ContributionTicker.scss
@@ -1,0 +1,81 @@
+// -----  gu-sass ----- //
+
+$progess-bar-height: 10px;
+
+.contributions-landing-ticker {
+  padding-bottom: $gu-h-spacing;
+}
+
+.contributions-landing-ticker__border {
+  border-top: 1px solid gu-colour(garnett-neutral-4);
+  margin: 0 -10px;
+}
+
+.contributions-landing-ticker--hidden {
+  visibility: hidden;
+  .contributions-landing-ticker__border {
+    visibility: visible;
+  }
+}
+
+.contributions-landing-ticker {
+  @include mq($until: tablet) {
+    display: none;
+  }
+
+  display: flex;
+  flex-direction: column;
+}
+
+.contributions-landing-ticker__values {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+}
+
+.contributions-landing-ticker__goal, .contributions-landing-ticker__so-far {
+  display: flex;
+  flex-direction: column;
+}
+
+.contributions-landing-ticker__so-far .contributions-landing-ticker__count {
+  font-family: $gu-titlepiece;
+  font-size: 32px;
+  font-weight: 800;
+  color: gu-colour(opinion-garnett-media-main-1);
+}
+
+.contributions-landing-ticker__goal {
+  align-items: flex-end;
+
+  .contributions-landing-ticker__count {
+    font-weight: bold;
+  }
+}
+
+.contributions-landing-ticker__label {
+  font-style: italic;
+}
+
+.contributions-landing-ticker__progress-bar {
+  position: relative;
+}
+
+.contributions-landing-ticker__progress {
+  width: calc(100%);
+  height: $progess-bar-height;
+  overflow: hidden;
+  background-color: #dcdcdc;
+  position: absolute;
+}
+
+.contributions-landing-ticker__filled-progress {
+  background-color: gu-colour(opinion-garnett-media-main-1);
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  transform: translateX(-100%);
+  transition: transform 3s cubic-bezier(.25, .55, .2, .85);
+}
